fix: seperated api calls and auth handlers
