+++ src/services/auth-service.js
@@ -1,21 +1,18 @@
 import axios from "axios";
 
-const loginHandler = async (loginCredentials, setAuthTokens, navigation) => {
+const loginUser = async (loginCredentials) => {
     try {
         const { data } = await axios.post("/api/auth/login", {
             email: loginCredentials.email,
             password: loginCredentials.password,
         });
-        const encodedToken = data.encodedToken;
-        localStorage.setItem("token", encodedToken);
-        setAuthTokens(encodedToken);
-        navigation("/");
+        return data;
     } catch (error) {
         console.log(error);
     }
 };
 
-const signupHandler = async (signupCredentials, setAuthTokens, navigation) => {
+const signupUser = async (signupCredentials) => {
     try {
         const { data } = await axios.post("/api/auth/signup", {
             firstName: signupCredentials.firstName,
@@ -23,23 +20,10 @@ const signupHandler = async (signupCredentials, setAuthTokens, navigation) => {
             email: signupCredentials.email,
             password: signupCredentials.password,
         });
-        const encodedToken = data.encodedToken;
-        localStorage.setItem("token", encodedToken);
-        setAuthTokens(encodedToken);
-        navigation("/");
+        return data;
     } catch (error) {
         console.log(error);
     }
 };
 
-const signoutHandler = async (removeAuthTokens, navigation) => {
-    try {
-        localStorage.removeItem("token");
-        removeAuthTokens();
-        navigation("/");
-    } catch (error) {
-        console.log(error);
-    }
-};
-
-export { loginHandler, signupHandler, signoutHandler };
+export { loginUser, signupUser };
